+++ src/main.ts
@@ -3,6 +3,7 @@ import App from './App.vue'
 import { vuetify } from './plugins/vuetify'
 import { router } from './router'
 import './style.css'
+import 'vuetify/styles'
 
 const app = createApp(App)